Create a common folder inside of the component folder to group everything that can be used in many situations and move the components that is used in the layout to the Layout folder
