--- src/components/Header/index.js
@@ -1,10 +1,0 @@
-import React from "react";
-import { FormattedMessage } from "react-intl";
-
-const Header = () => (
-  <>
-    <div id="header" />
-  </>
-);
-
-export default Header;
