+++ src/components/Layout/index.js
@@ -1,8 +1,8 @@
 import React from "react";
 import PropTypes from "prop-types";
-import Navbar from "../Navbar";
-import Footer from "../Footer";
-import Sidebar from "../SideBar/Sidebar";
+import Navbar from "./Navbar";
+import Sidebar from "./SideBar";
+import Footer from "./Footer";
 
 const Layout = ({ children }) => (
   <div className="flex flex-no-wrap h-full min-h-screen">
