+++ src/pages/projetos/index.js
@@ -6,7 +6,7 @@ import Link from "next/link";
 import { useRouter } from "next/router";
 import Layout from "../../components/Layout";
 import usePreviousPath from "../../contexts/previousPath";
-import Breadcrumbs from "../../components/Breadcrumbs";
+import Breadcrumbs from "../../components/common/Breadcrumbs";
 
 const GreenCircle = styled.div`
   width: 20px;
